TST: Fixed more tests (4 still failing) MAINT: Moved `AnalysisController` into separate file
