+++ topobank/manager/tests/test_search.py
@@ -157,7 +157,7 @@ def test_surface_search_with_request_factory(user_three_surfaces_four_topographi
             'description': '',
             'folder': True,
             'key': f'surface-{surface1.pk}',
-            'label': surface1.name,
+            'label': surface1.label,
             'name': surface1.name,
             'id': surface1.pk,
             'publication_authors': '',
@@ -236,7 +236,7 @@ def test_surface_search_with_request_factory(user_three_surfaces_four_topographi
             'description': '',
             'folder': True,
             'key': f'surface-{surface2.pk}',
-            'label': surface2.name,
+            'label': surface2.label,
             'name': surface2.name,
             'id': surface2.pk,
             'publication_authors': '',
@@ -269,7 +269,7 @@ def test_surface_search_with_request_factory(user_three_surfaces_four_topographi
             'description': '',
             'folder': True,
             'key': f'surface-{surface3.pk}',
-            'label': surface3.name,
+            'label': surface3.label,
             'name': surface3.name,
             'id': surface3.pk,
             'publication_authors': '',
@@ -344,7 +344,7 @@ def test_surface_search_with_request_factory(user_three_surfaces_four_topographi
             'description': '',
             'folder': True,
             'key': f'surface-{surface2.pk}',
-            'label': surface2.name,
+            'label': surface2.label,
             'name': surface2.name,
             'id': surface2.pk,
             'publication_authors': '',
@@ -443,7 +443,7 @@ def test_surface_search_with_request_factory(user_three_surfaces_four_topographi
             'description': '',
             'folder': True,
             'key': f'surface-{surface1.pk}',
-            'label': surface1.name,
+            'label': surface1.label,
             'name': surface1.name,
             'id': surface1.pk,
             'publication_authors': '',
@@ -923,7 +923,7 @@ def test_tag_search_with_request_factory(user_three_surfaces_four_topographies):
                     'publication_date': '',
                     'key': f"surface-{surface4.pk}",
                     'folder': False,
-                    'label': surface4.name,
+                    'label': surface4.label,
                     'name': surface4.name,
                     'selected': False,
                     'children': [],
